+++ react/ProductQuantity.tsx
@@ -5,12 +5,7 @@ import { useProductDispatch } from 'vtex.product-context/ProductDispatchContext'
 import BaseProductQuantity, { Props } from './components/BaseProductQuantity'
 
 const ProductQuantity: StorefrontFunctionComponent<Props> = props => {
-  const {
-    warningQuantityThreshold,
-    showLabel,
-    showMeasurementUnit,
-    size,
-  } = props
+  const { warningQuantityThreshold, showLabel, size } = props
   const { selectedItem, selectedQuantity } = useProduct()
   const dispatch = useProductDispatch()
   return (
@@ -18,7 +13,6 @@ const ProductQuantity: StorefrontFunctionComponent<Props> = props => {
       size={size}
       dispatch={dispatch}
       showLabel={showLabel}
-      showMeasurementUnit={showMeasurementUnit}
       selectedItem={selectedItem}
       selectedQuantity={selectedQuantity}
       warningQuantityThreshold={warningQuantityThreshold}