+++ react/ProductSummaryQuantity.tsx
@@ -8,12 +8,7 @@ import BaseProductQuantity, { Props } from './components/BaseProductQuantity'
 const CSS_HANDLES = ['summaryContainer'] as const
 
 const ProductSummaryQuantity: StorefrontFunctionComponent<Props> = props => {
-  const {
-    warningQuantityThreshold,
-    showLabel,
-    showMeasurementUnit,
-    size,
-  } = props
+  const { warningQuantityThreshold, showLabel, size } = props
   const handles = useCssHandles(CSS_HANDLES)
   const { selectedItem, selectedQuantity } = useProduct()
   const dispatch = useProductDispatch()
@@ -33,7 +28,6 @@ const ProductSummaryQuantity: StorefrontFunctionComponent<Props> = props => {
         size={size}
         dispatch={dispatch}
         showLabel={showLabel}
-        showMeasurementUnit={showMeasurementUnit}
         selectedItem={selectedItem}
         selectedQuantity={selectedQuantity}
         warningQuantityThreshold={warningQuantityThreshold}